Removing context abstraction layer.
This layer is no longer useful now that game development is done through Lua. I might come up with another wrapper later if I see a need for it but it would have to be generated automatically using a script and the information found in Bindings.h.

Also changed the position for the game window. The version of SDL I am using doesn't center the window very well in Windows when high a DPI monitor is used.